Reorganized based on updated folder structure, added DialectQuidemTest
