+++ parsing/dialects/default/src/test/resources/sql/agg.iq
