+++ parsing/dialects/default/src/test/resources/sql/join.iq
