+++ parsing/dialects/default/src/test/resources/sql/lateral.iq
