+++ parsing/dialects/default/src/test/resources/sql/materialized_view.iq
