+++ parsing/dialects/default/src/test/resources/sql/misc.iq
