+++ parsing/dialects/default/src/test/resources/sql/outer.iq
