+++ parsing/dialects/default/src/test/resources/sql/scalar.iq
