+++ parsing/dialects/default/src/test/resources/sql/stream.iq
