+++ parsing/dialects/default/src/test/resources/sql/struct.iq
