+++ parsing/dialects/default/src/test/resources/sql/sub-query.iq
