+++ parsing/dialects/default/src/test/resources/sql/table.iq
