+++ parsing/dialects/default/src/test/resources/sql/table_as.iq
