+++ parsing/dialects/default/src/test/resources/sql/view.iq
