+++ parsing/dialects/hive/src/test/resources/sql/select.iq
