+++ parsing/dialects/mysql/src/test/resources/sql/select.iq
