+++ parsing/dialects/postgresqlBase/dialects/redshift/src/test/resources/sql/redshift.iq
